// src/modules/payments/enums/payment-status.enum.ts
export enum PaymentStatus {
  PENDING = 'PENDING',
  INITIATED = 'INITIATED',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
  CANCELLED = 'CANCELLED',
  REFUNDED = 'REFUNDED',
}

export enum PaymentFailureReason {
  INSUFFICIENT_FUNDS = 'INSUFFICIENT_FUNDS',
  INVALID_PHONE = 'INVALID_PHONE',
  TIMEOUT = 'TIMEOUT',
  USER_CANCELLED = 'USER_CANCELLED',
  MERCHANT_INACTIVE = 'MERCHANT_INACTIVE',
  SYSTEM_ERROR = 'SYSTEM_ERROR',
  DARAJA_ERROR = 'DARAJA_ERROR',
}